changes on pom.xml & users java file

+++ srm-rest/src/main/java/com/srm/services/entity/Users.java
@@ -9,17 +9,17 @@ import javax.persistence.Table;
 
 @Entity
 @Table(name="USERS")
-public class Users implements Serializable{
-	
+public class Users implements Serializable {
+
 	@Id
 	@Column(name = "USER_ID")
-	private Integer userId;
+	private Long userId;
 
-	public Integer getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(Integer userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
